chore: extend CI workflow with rust-clippy analysis
The changes introduce rust-clippy analysis to the existing CI workflow. A new job named 'rust-clippy-analyze' has been added to the workflow which needs the 'build-and-test' job. This job runs after the project compilation and tests, it also produces a SARIF report about the code checks performed and upload it to GitHub