feat: more game logic changes with sprite and text generation
This commit incorporates a significant update to the game logic and visual elements of the Rust Minesweeper clone. Mine sprites are used and text containing number of mines neighboring the tiles are added when generating the board. By adding new `mine`, `mine_neighbor`, and `uncover` structs, we have implemented the basic components of the game that will be used within the Bevy game engine. Default window resolution was changed to 800 x 800