feat: add new tile generation and components (#5)
* chore: add Pixelify Sans font and associated documentation

This commit adds the Pixelify Sans font to the project, in both variable and static versions. A README is included, providing an overview of the font's features and uses. The license for the font, the SIL Open Font License, is also included in the package

* chore: add initial sprites for mines and flags

* chore: remove Pixelify_Sans font, add Symtext font

This commit removes the files associated with the Pixelify_Sans font from the `assets` directory. In replacement, a new files for the Symtext font are added. This is because Pixelify_Sans number `2` could be mistaken with `S` character easily

* feat: more game logic changes with sprite and text generation

This commit incorporates a significant update to the game logic and visual elements of the Rust Minesweeper clone. Mine sprites are used and text containing number of mines neighboring the tiles are added when generating the board. By adding new `mine`, `mine_neighbor`, and `uncover` structs, we have implemented the basic components of the game that will be used within the Bevy game engine. Default window resolution was changed to 800 x 800