feat: uncovering tiles (#7)
* chore: update README.md

The README.md was updated to include a link to a tutorial for a Rust-based Minesweeper clone. The updated project utilises the latest Bevy 0.12 for the tutorial adaptation

* chore: add JetBrains AI assistant mention in README.md

Included a new line in the README.md mentioning use of JetBrains AI assistant for generating documentation, code, and commit messages. This additional information gives users an idea of tools used in development

* chore: README.md TOC and Features section updates

This update includes the addition of a detailed table of contents in the README.md file for better navigation and updates to the Features section

* chore: add covered tiles tracking to Board struct

A new 'covered_tiles' field has been added to the 'Board' struct allowing to map coordinates to uncovered tile entities. Relevant methods to manipulate and retrieve these tiles have also been included. Changes include modifications in the 'board_plugin' for the spawning logic of the covered tiles.

* chore: add TileTriggerEvent

Added a new event called `TileTriggerEvent` to handle tile interaction upon left mouse button (LMB) clicks. The `TileTriggerEvent` is being passed with the tile coordinates which allows further interaction with the tile.

* feat: rename `board` to `board_plugin` and add `uncover` system

The `board` module has been renamed to `board_plugin`. A new `uncover` system is added to handle triggering and processing of uncovering tiles. This system interacts with `BoardPlugin` to add functionality for uncovering tiles upon receiving a `TileTriggerEvent`.