feat: increase starting `mine_count`, modify board color, and implement `safe_start` (#8)
In the `BoardOptions`, the `mine_count` was increased from 40 to 60, and the `safe_start_enabled` option was added and set to true. The `color` in the board `SpriteBundle` was changed from `ANTIQUE_WHITE` to `BLACK`. New logic was introduced to handle the `safe_start` function where the first uncovered tile will always be safe.